requirements.txt
altgraph==0.17.2
art==5.7
asgiref==3.5.0
async-generator==1.10
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.11.1
Brotli==1.0.9
build==0.10.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cleo==2.0.1
click==8.1.3
colorama==0.4.5
crashtest==0.4.1
cryptography==36.0.2
dash==2.5.1
dash-bootstrap-components==1.1.0
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
DateTime==4.4
decorator==5.1.1
distlib==0.3.4
dj-database-url==1.2.0
Django==4.0.3
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
docopt==0.6.2
docutils==0.18.1
dulwich==0.21.3
et-xmlfile==1.1.0
exceptiongroup==1.1.0
filelock==3.9.1
Flask==2.1.2
Flask-Compress==1.12
future==0.18.2
gevent==21.12.0
glob2==0.7
google==3.0.0
google-api-core==2.10.1
google-api-python-client==2.72.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
greenlet==1.1.2
h11==0.14.0
html5lib==1.1
httplib2==0.21.0
idna==3.3
importlib-metadata==6.0.0
installer==0.6.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.0.3
jsonschema==4.17.3
kami-logging==0.2.0
keyring==23.13.1
Kivy==2.1.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
kivymd==1.1.1
localflavor==1.9
lockfile==0.12.2
lxml==4.8.0
MarkupSafe==2.1.1
more-itertools==9.1.0
msgpack==1.0.5
num2words==0.5.10
numpy==1.22.4
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.0.10
outcome==1.2.0
packaging==23.0
pandas==1.4.2
passlib==1.7.4
pefile==2022.5.30
pexpect==4.8.0
Pillow==9.0.1
pkginfo==1.9.6
platformdirs==2.6.2
plotly==5.8.2
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
polib==1.1.1
protobuf==4.21.6
psutil==5.9.0
psycopg2==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.6
Pygments==2.13.0
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyPDF2==1.26.0
pypiwin32==223
pyproject_hooks==1.0.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2021.3
pywin32==303
pywin32-ctypes==0.2.0
rapidfuzz==2.13.7
reportlab==3.6.8
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
selenium==4.8.2
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tenacity==8.0.1
tomli==2.0.1
tomlkit==0.11.6
trio==0.22.0
trio-websocket==0.9.2
trove-classifiers==2023.3.9
typing_extensions==4.5.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.9
urwid==2.1.2
virtualenv==20.13.3
webencodings==0.5.1
Werkzeug==2.1.2
whitenoise==6.3.0
wkhtmltopdf==0.2
wsproto==1.2.0
xlrd==2.0.1
xmltodict==0.13.0
zipp==3.15.0
zope.event==4.5.0
zope.interface==5.4.0
